fix(testing): fix tests for Windows (#165)
In this PR, I identified and fixed various issue while running tests in
the Windows environment:

- shell needs to be specified to run programs in command shell on
Windows, as described
[here](https://stackoverflow.com/questions/60386867/node-spawn-child-process-not-working-in-windows)
- path notation on Windows is different
- JSON formatting is different for UNIX and Windows
- writing into files replaced by echo command which works seamlessly on
Windows
- glob patterns require extra quotation marks for argument expansion on
Linux
- recreating tmp folder caused issues for me so I partially fixed it via
adding the recursive option

Note: This is the same scope as #131. Unfortunately, merging changes
from the test refactor did not go well so I started off a clean branch
and re-applied the changes.