ci: add custom `release` type to commitlint configuration (#309)
Now, `commitlint` will accept `release` as a type for commit messages
along with the previously existing types, allowing more specific
categorization of release commits. This change doesn't affect any
existing functionality but extends the types that can be used for commit
messages.

`release` was added for release actions after merging in the main.