fix(plugin-eslint): eslint startline can be zero (#563)
In case tsconfig does not contain correct setup, e.g.,
"strictNullChecks": true, eslint throws warnings for line "zero". The
current schema does not show the warning, it just throws an internal
error. This fix allows for warning on line zero to be collected for the
report.

Fixes issues such as:
  ```
{
    "code": "too_small",
    "minimum": 0,
    "type": "number",
    "inclusive": false,
    "exact": false,
    "message": "Number must be greater than 0",
    "path": [
      117,
      "details",
      "issues",
      3,
      "source",
      "position",
      "startLine"
    ]
  },

```

for issues like 
`0:1  warning  This rule requires the `strictNullChecks` compiler option to be turned on to function correctly  @typescript-eslint/prefer-nullish-coalescing`

---------

Co-authored-by: Markus.Nissl <Markus.Nissl@Entaingroup.com>