test(nx-plugin-e2e): remove flaky "--verbose" substring assertion
Fails when run in Nx Cloud on a commit which happens to have
"--verbose" in the title (such as this one or parent).

The verbose argument handling is already tested in
executor.unit.test.ts, so it's not necessary to test such
a non-critical feature again in E2E tests.